==> movie/code/4. naver_star.ipynb <==
# 4. 네이버 OPEN API

네이버 데이터 OPEN API를 사용하여 영화별 별점과 검색량 자료를 CSV파일로 저장

## 1) 네이버 별점

네이버 영화에서 bs4를 이용하여 크롤링으로 관람객 평점(네이버영화예매자), 네티즌 평점, 기자평론가 평점을 수집

## 2) 네이버 데이터랩

영화 이름과 기간을 입력해 해당기간에 해당영화를 얼마나 검색했는지 확인


.
.
### 4.1 네이버 별점

### 1) requests 모듈로 url 내용을 가져오고 BeautifulSoup로 내용 가공한다.

from bs4 import BeautifulSoup
from urllib.request import urlopen
import requests
import pandas as pd
from tqdm import tqdm
from tqdm import tqdm_notebook
from datetime import date
import glob
import multiprocessing
import os

movie_info_path='/home/ubuntu/miniii_su/movie/'

start_date = date(2017,1,1)
end_date = date(2018,2,28)

def naver_movie_daily_boxoffice(single_date):
    j=0
    movie_name=[]
    movie_point=[]
    targetDt = str(single_date)[0:4]+str(single_date)[5:7]+str(single_date)[8:10]
    url = "http://movie.naver.com/movie/sdb/rank/rmovie.nhn?sel=cnt&date="+targetDt
    res = requests.get(url)
    soup = BeautifulSoup(res.content, "html.parser")
    b = soup.find_all('div','tit3') # 전체 영화의 제목 뽑아옴
        
    for n in range(0, len(b)):
        movie_name.append(b[n].find('a').text)  # 영화명, 평점 뽑아옴
        movie_point.append(b[n].find('a')['href'][28:])
        
    dff = list(zip(movie_name,movie_point))
    dff = pd.DataFrame(dff)
    dff.to_csv(movie_info_path+'tmp_data/'+targetDt+'.csv',index=False)

cpu = os.cpu_count() # cpu 사용환경에 따라서 맞추기 위함
date_list = pd.date_range(start_date, end_date)
pool = multiprocessing.Pool(processes=cpu) # 여러개 동시에 돌림
pool.map(naver_movie_daily_boxoffice,date_list)
pool.close()
pool.join()

### 2) 날짜별로 수집된 csv 파일을 하나로 합치기

tmp = glob.glob(movie_info_path+'tmp_data/*.csv')
dff = pd.DataFrame()
for i in tmp:
    df = pd.read_csv(i)
    dff = pd.concat([df,dff])
dff.to_csv(movie_info_path+'data/naver_boxoffice.csv',index=False)

### 3) 하나로 합쳐진 파일을 만들었던 각각의 날짜별 파일 삭제

# tmp_data = glob.glob(movie_info_path+'tmp_data/*.csv')
# for data in tmp_data:
#     os.remove(data)

dff = pd.read_csv(movie_info_path+'data/naver_boxoffice.csv')
dff = dff.drop_duplicates()
dff2 = pd.DataFrame(columns=['2','3','4'])
dff = pd.concat([dff,dff2])
dff.to_csv(movie_info_path+'data/naver_boxoffice.csv',index=False)

import numpy as np
star1=[]
star2=[]
star3=[]
list1=[]
list2=[]


num1=df['1']

def naver_movie_rating(code):
    df = pd.read_csv(movie_info_path+'data/naver_boxoffice.csv')
    idx = np.where(df['1']==code)
    
    url1 = "https://movie.naver.com/movie/bi/mi/basic.nhn?code="
    page1 = urlopen(url1+str(code)[:-2])
    soup1 = BeautifulSoup(page1, "html.parser")
    b = soup1.find_all('span','st_on')
    
    if(len(b)<3):
        df['2'][idx[0][0]] = 0
        df['3'][idx[0][0]] = 0
        df['4'][idx[0][0]] = 0
    else:
        if(b[2]['style'][6:-1] == '0'):
            df['2'][idx[0][0]] = round(float(b[0]['style'][6:-1])/10,2) # idx가 array로 나옴
            df['3'][idx[0][0]] = 0
            df['4'][idx[0][0]] = round(float(b[1]['style'][6:-1])/10,2)
        else:
            df['2'][idx[0][0]] = round(float(b[0]['style'][6:-1])/10,2)
            df['3'][idx[0][0]] = round(float(b[1]['style'][6:-1])/10,2)
            df['4'][idx[0][0]] = round(float(b[2]['style'][6:-1])/10,2)

    df.to_csv(movie_info_path+'data/naver_boxoffice.csv',index=False)

### 4) naver_movie_rating 사용자 정의함수를 이용하여 별점 수집

df = pd.read_csv(movie_info_path+'data/naver_boxoffice.csv')
for code in tqdm(df['1']):
    naver_movie_rating(code)

import pandas as pd
df = pd.read_csv(movie_info_path+'data/naver_boxoffice.csv')
df2 = df.drop_duplicates()
df2 = df2.reset_index(drop=True)
df2 = df2.rename(columns={df2.columns[0]:'movieNm',df2.columns[1]:'naver_movieCd', df2.columns[2]:'ntz_see_score', df2.columns[3]:'spc_score', df2.columns[4]:'ntz_score'})
df2.to_csv(movie_info_path+'data/naver_boxoffice.csv',index=False)
