==> movie/code/5. merge.ipynb <==
# 5. 데이터 전처리
### 1) kobis와 네이버 영화에서 수집한 데이터 중
### kobis 영화코드와 네이버 영화코드를 단일화 작업 후 
### 분석에 필요한 데이터만 추출

import pandas as pd
import numpy as np

movie_info_path='/home/ubuntu/miniii_su/movie/'

df1 = pd.read_csv(movie_info_path+'data/movie_code_add_director.csv')
df2 = pd.read_csv(movie_info_path+'data/naver_boxoffice.csv')
df1 = df1.drop('Unnamed: 0',1)
df1 = df1.drop('Rate',1)
df3 = pd.merge(df1,df2,how='left',on='movieNm')

for i in range(len(df3['ntz_see_score'])):
    if(np.isnan(df3['ntz_see_score'][i])):
#         print(str(i)+" NaN")
        name = df3['movieNm'][i].replace(' ','')
#         print(name)
        for j in range(len(df2)):
            df2_name = df2['movieNm'][j].replace(' ','')
#             print(df2_name)
            if(df2_name[0:8] == name[0:8]):
                df3['naver_movieCd'][i] = df2['naver_movieCd'][j]
                df3['ntz_see_score'][i] = df2['ntz_see_score'][j]
                df3['spc_score'][i] = df2['spc_score'][j]
                df3['ntz_score'][i] = df2['ntz_score'][j]
                if(df2_name[0:6] == name[0:6]):
                    df3['naver_movieCd'][i] = df2['naver_movieCd'][j]
                    df3['ntz_see_score'][i] = df2['ntz_see_score'][j]
                    df3['spc_score'][i] = df2['spc_score'][j]
                    df3['ntz_score'][i] = df2['ntz_score'][j]
                    if(df2_name[0:4] == name[0:4]):     
                        df3['naver_movieCd'][i] = df2['naver_movieCd'][j]
                        df3['ntz_see_score'][i] = df2['ntz_see_score'][j]
                        df3['spc_score'][i] = df2['spc_score'][j]
                        df3['ntz_score'][i] = df2['ntz_score'][j]            
    elif(np.isnan(df3['naver_movieCd'][i])):
        df3['naver_movieCd'][i] = 0
        df3['ntz_see_score'][i] = 0
        df3['spc_score'][i] = 0
        df3['ntz_score'][i] = 0

df3 = df3.dropna()
df3 = df3.reset_index(drop=True)
df3

df3.to_csv(movie_info_path+'data/naver_movie_all.csv',index=False)
